convert to double

+++ array.cpp
@@ -3,20 +3,20 @@
 Array::Array(int n){
 	int i;
 	this->n=n;
-	item=new float[n];
+	item=new double[n];
 	for(i=0;i<n;i++){
 		item[i]=0;
 	}
 }
-Array::Array(const float *x,const int n){
+Array::Array(const double *x,const int n){
 	int i;
 	this->n=n;
-	this->item=new float[n];
+	this->item=new double[n];
 	if(x){
 		for(i=0;i<n;i++){
 			this->item[i]=x[i];
 		}
-//			memcpy(this->item,x,n*sizeof(float));
+//			memcpy(this->item,x,n*sizeof(double));
 	}
 }
 Array::~Array(){
@@ -24,7 +24,7 @@ Array::~Array(){
 }
 void Array::print(){
 	int i;
-	float *x;
+	double *x;
 	x=this->item;
 	for(i=0;i<this->n;i++){
 		printf("[%i]%.02f\t",i,*x++);
@@ -34,6 +34,6 @@ void Array::print(){
 void Array::rand(){
 	int i;
 	for(i=0;i<this->n;i++){
-		this->item[i]=(random()-(RAND_MAX/2))*2.0/((float)RAND_MAX)/((float)RANDSCALING);
+		this->item[i]=(random()-(RAND_MAX/2))*2.0/((double)RAND_MAX)/((double)RANDSCALING);
 	}
 }
