+++ matrix.cpp
@@ -4,7 +4,7 @@ Matrix::Matrix(int m,int n){
 	int i;
 	this->m=m;
 	this->n=n;
-	item=new float[m*n];
+	item=new double[m*n];
 	for(i=0;i<m*n;i++){
 		item[i]=0;
 	}
@@ -12,20 +12,20 @@ Matrix::Matrix(int m,int n){
 Matrix::~Matrix(){
 	delete item;
 }
-float Matrix::atIndex(int i,int j) const{
+double Matrix::atIndex(int i,int j) const{
 	//assert(((i+1)*(j+1))<=(n*m));
 	return item[i*n+j];
 }
-float Matrix::atIndex(int i,int j){
+double Matrix::atIndex(int i,int j){
 	//assert(((i+1)*(j+1))<=(n*m));
 	return item[i*n+j];
 }
 void Matrix::rand(){
 	int i,j;
-	float *p=item;
+	double *p=item;
 	for(i=0;i<this->m;i++){
 		for(j=0;j<this->n;j++){
-			*p++=(random()-(RAND_MAX/2))*2.0/((float)RAND_MAX)/((float)RANDSCALING);
+			*p++=(random()-(RAND_MAX/2))*2.0/((double)RAND_MAX)/((double)RANDSCALING);
 		}
 	}
 }
